+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Addressfinder 1.15.0 (March 2025) #
 
-* Automatically skip empty strings within Bulk verification
+* Automatically skip empty strings within Batch verification
+* Mark unverified addressses as false within Batch verification
 
 # Addressfinder 1.14.0 (February 2025) #